+++ src/data_sync/config.py
@@ -1,7 +1,5 @@
 """Configuration details for sync jobs"""
 
-from __future__ import annotations
-
 from dataclasses import dataclass
 from pathlib import Path
