+++ src/fetch/orderbook.py
@@ -16,7 +16,7 @@ from sqlalchemy.engine import Engine
 from src.logger import set_log
 from src.models.block_range import BlockRange
 from src.utils.query_file import open_query
-from src.dbt.common import node_suffix
+from src.data_sync.common import node_suffix
 
 log = set_log(__name__)
